Added Factory for Pages
Added a factory for page generation. It uses the new "available_pages"
from the Page to figure out what it can/can't generate and then creates
them. Redid bin/dotgrid and the generator to just pass in the page types
that are needed.

Still need a) more tests and b) redo the README + "base" (still need to
rename that or something).